First commit
We have preplaced control points and moving zombies. Rather than
calculate velocity for every object based on the control points, I
decided to save a 2D array of velocities based on the control points,
and refresh that array every time a control point is placed or removed.
When a zombie asks for a velocity at its position it will receive the
weighted average of the four velocities around it, with weightings
proportional to closeness. This will be much more scalable.

There are some odd bugs with the movement that I'm trying to work out
still, and obviously no mouse controls have been configured.