Control point system adjustments
So every frame each zombie needs a movement direction, which is
calculated based on the positions of the control points. For efficiency
reasons, we can't have every zombie doing that calculation every turn,
so instead the game calculates movement directions at particular points
throughout the map, and when a zombie asks for a direction it gives it
the average of the nearest four. The pre-calculated direction vectors
make up the "vector field," a literal vector field projected onto the
game world.

Previously, that field started at 0,0 and went as far as it was told to.
With this commit, the vector field is now tied to a map object
(specifically the box collider on the map object). We can move and
resize the map however we want and the vector field will follow. We can
also define the granularity of the vector field for performance
optimization.