Zombie movement change, control point sprite scaling, lots of optimizations
Zombies will no longer move exactly as they should, but will instead
move with pseudo-acceleration. This causes them to swarm around control
points rather than stop in front of them. It's pretty cool looking.

Control points now scale according to camera zoom so they can be easily
seen and grabbed when zoomed out.

Shit tons of optimizations, the biggest of which is splitting vector
field recalculations over multiple frames. The game now runs perfectly
with around 280 zombies and 30 humans, and with a 500 by 500 map, which
can easily hold ten times that many objects. I have a number of other
optimizations in mind that will make it even more scalable, which is
good because the human AI is gonna be expensive.