Starting to implement range filtering for unit checks
Currently every zombie looks at every human every frame to see which are
in range and which is the closest. I've begun implementing changes to
get them to only look at those that are near them on the map. I have to
stop working now, so this commit doesn't compile, but I wanted to save
my work.

For future me:
MapInfo is complete. What isn't complete is converting the object
dictionary from a dictionary of lists to a dictionary of 2d arrays of
lists. The 2d array represents the GO tracker's map squares.