Added map square optimizations
The game map is now broken up into squares. When a game object requests
a list of nearby objects from the GO tracker, instead of giving back a
list of every game object, the tracker will give back a list of objects
that are within only the surrounding map squares.

Because of this we're now running smoothly with ~1700 zombies, but I
think we'll need at least 5000 to get the sort of experience we're going
for.