Setting the stage for human AI
Created prefabs for all three resources and for camps, and set up all
the class info to support harvesting resources, depositing resources in
camps, requesting resources from camps, and death from resource
deprivation. Also added code for humans to look for camps/resources, and
exchange information with each other. All that's really left is to add
the decision-making code and we'll have a mostly functional game.