Human wander and evade, zombie drift, bug fixes and optimizations
Humans will wander if they can't locate a camp and will run from
zombies.

I also fixed the issue that was causing large numbers of humans to do
nothing - I generate seed numbers for things that don't happen every
frame (which is everything, basically) so that all the units that do
them do them on separate frames, but for lots of those things I had the
range for the seed one to high, which meant sometimes a seed would be
generated that was never called, preventing the action from ever
happening. There are no humans standing around anymore.

Zombies will no longer come to a stop if all control points are removed.
If you let them swarm a control point for a while you can release it for
a cool explosion effect, it's pretty cool.

More optimizations, mostly of the human AI. The biggest one is that
greets are now limited to a certain number per turn. With wander
implemented there just isn't a need for the spread of information to be
instantaneous, so there's no need for constant greets.

I was able to run it with 2400 humans and 1700 zombies. We're getting
there.