Human attack
Humans will tag the closest zombie if it's in their attack range,
killing it. They can't tag another zombie for another few frames after
that. The tag is a guaranteed kill, but the humans still don't really
stand a chance against an equal number of zombies, which is strange.
Movement speed is probably a large factor.