Added walking loops + tweaked scale
Found out how to make the sprites animate. It will be more complex in
the final one obviously, but for now it just loops through the 4 frames.
The fact that they're all in sync right now is annoying me, but that
will have to be addressed later.

I also tweaked some of the scales so they seemed to match the movement
better. Obviously all of that is still very workable.