Added a start screen
Now we have a start screen that has two buttons (one to start the game,
one to quit the game).

I also added a line to the camera controls that just makes it so ESC
quits the game. It's certainly a crude solution, but it's a good
temporary fix for being trapped in the game if it's full screen.