Changed visuals/animation
I changed the coloring on the sprites so that the zombies are more
distinguished from the humans. Hopefully that comes across better now.

I also tweaked how the sprite sheets are being initially interpreted to
attempt to improve performance but honestly I don't think it did much.
The only solution I can think now is to reduce the amount of poses from
4 to 2. The animations will be a bit more "blocky" but it might be worth
it for performance sake.