Adding stats script, fixing issue with control points
New Stats script. Really simple, just stores stat values and provides
methods to read and alter them. The actual functionality of the stats is
left to whatever other script uses them. So far skill's functionality
has been partially implemented for humans.

Also discovered that adding dead zones and active radii to control
points broke them. I was checking whether the active point in the vector
field is within the active range of the control point, and if it wasn't
I just didn't add its vector to the array used to calculate the final
direction of the vector field vector. However, the array itself still
had a length of greater than one, so scaling was still being applied,
causing the single non-zero vector to be scaled down to a zero vector.
The result was that if there were only two control points, any position
that didn't fall within /both/ of their active ranges was given a zero
vector. I'm not sure whether the same happened with more than two
control points, but I don't think it would've.

The solution was to use a list instead of an array so that the count
would accurately reflect how many control points were considered to be
active for each point in the vector field.