Added new assets and animations
I made the 2 frame sprite sheets, had a problem using them if they were
on a combined sprite sheet so I just split them into two.

Added an extra resource for the fuel and food and got them to randomly
initialize to one of the two available. It's hardcoded so the value
would need to be changed every time we add a new option, but it works.

I attempted to get the "turning" working properly but failed. I left my
attempt lines commented out for reference, but I couldn't get the sprite
and animation controller loading properly. Instead I changed the color
values that they change to to make changed humans more visually obvious.