Random world generation
All pre-placed objects have been removed. Instead, a certain number of
each resource and a certain number of humans will be placed randomly
when the game starts. A number of zombies will be placed as well, though
they will be placed around a random point. The camera will be centered
on the zombies, so the player doesn't have to hunt around for them.