Control point bounding
Zombies will now prefer control points to humans if the ratio between
the distance to the nearest control point and the distance to the
nearest human is less than a configurable amount. I have it set to 1 for
now, which means zombies will prefer control points if they're closer to
the control point than they are to their target.

Under the hood, this required me to give the CPManager another field to
track in addition to the vector field. The new field represents
distances from control points, and is populated and drawn from in
exactly the same way as the vector field.