Map bounds checking fix
Fixing an error (I think) with the bounds checking at the edge of the
map. Drawing from the CP vector field checks the nearest four points, so
it isn't enough that the zombie's grid position be within the array, it
needs to have an extra element on top of that or there won't be enough
vectors to draw from. The vectors are bottom-left anchored in their
cells, so this was only an issue on the top and right sides of the map.

A really non-intuitive off-by-one error, basically.