Randomized map tiling, altitude parsing pt1
Map tiles will now be chosen at random from the full set of 11 (I'm
ignoring the extra empty tiles), and will be placed with random
rotations (rounded to multiples of 90).

Additionally, the groundwork is done for parsing altitude information
from the tiles based on color. In Start, MapInfo will create an altitude
field (similar to the CPM's vector field, but with floats instead of
vectors) and populate it by sampling pixel data from across the map and
running the color information through a conversion method to turn it
into an altitude. All that's left to write is the conversion method,
which will presumably check the color against a list of pre-specified
color-altitude pairs and return the closest match.

One potential problem is that I'm pretty sure Texture.GetPixel() doesn't
correct for object rotation, so I'll probably have to save the random
rotation values for each sprite and account for the rotations myself.
But I have no way to test that yet, so I'll with it when it happens.