Blue repulsor points
Repulsor points are now blue. Because start methods aren't called
immediately on object creation, the reference to the CP's sprite hadn't
been created by the time the CPM called the set color method. The hacky
solution is to call the set color method as a coroutine and have it skip
a frame as long as the sprite reference was null. This allowed it to
basically wait until it had the sprite renderer to change the color. It
works, though I think there's technically still a single frame where the
point is red instead of blue.