Better water node generation
Threw out most of the code from the previous commit; rather than placing
water nodes at the centroid of each body of water and handling
visibility via trickery in the GOT and GetHarvestRange, we're just gonna
place nodes around each body of water and have them act like regular
nodes (though they'll be invisible). There's less to fuck up this way,
and it'll allow water nodes to act the same way they do already.

To explain the algorithm a bit:
We iterate through the altitude field, adding water points that are
adjacent to non-water points to a list. Then we iterate through that
list, checking each point against the list of already chosen node
locations. If the border point is far enough away from all the node
locations it becomes a node location. Finally, we pass the node
locations to the GOT so it can create stuff, same as the centroid
method.

So water is fully functional now.