Human AI fix
Humans were getting stuck near camps. This was because when the branch
of the decision making code where the nearest camp is walkable but the
nearest needed resource is not was reached, targetPos was being set
while targetPosTolerance was not. This caused the AI to correctly
navigate to the camp, but not to get close enough that they could
actually interact with the camp. If the default tolerance had been set
lower I might not even have noticed this. The solution was, obviously,
to set targetPosTolerance to the correct value at this point in the
code.

I also added another measure to possibly fix the human counter. I'm
still thinking the issue is that humans are being destroyed several
turns after marking themselves for destruction, so they're triggering
the ReportDeath call multiple times per death. We'll see if this is
actually the case.