Fog of war, toggle for density maps, camp strength/ammo, respawnable resources
This is a big one.

Fog of war is implemented. It's very simple, you can only see what your
units see, and everything else is black. No terrain memory, since I
haven't figured out how to occlude based on transparency, and probably
won't have time to. This honestly adds a lot to the experience, even if
it doesn't change the gameplay much.

Density map visuals can be toggled with E. Pretty straightforward.

Camp strength is a thing now. Basically, camps now have a strength
pseudo-stat that increases as resources are drawn from the camp, and
this pseudo-stat is scaled into a multiplier between 1 and a
configurable number (6 right now, I think). This number is inherited by
humans who interact with the camp, and serves as an attack range
multiplier if and only if they have ammo on them. No ammo and they
revert to normal attack range. Humans can carry three times as much ammo
as everything else right now, so nine total, meaning they get nine
attacks at the boosted range. A camp creates two ammo for every one of a
resource that gets deposited.

When a resource node is depleted, it will spawn a new resource node of
the same type nearby (technically it moves itself there and resets its
stock). The range is configurable, and it will only respawn at sea
level.