Fixing mesh bounding errors
For whatever reason, certain maps sizes caused the density maps and fog
of war to be frustum occluded at certain edges of the map. I've fixed
this by setting the bounds on all new meshes to the correct size for the
map.

Also working on limiting the camera to the map, so players won't be able
to see things that are off the map. Because the fog of war only extends
to the edge of the map, this will be necessary.