Clickable minimap, camera bounding, node respawn bounding, minimap control point graphic, various bug fixes
Click and drag on the minimap to move the main camera.

Main camera is constrained to the map.

Nodes will no longer be able to respawn outside the map.

Control points will be drawn on the minimap at a constant scale that's
large enough for the player to see. This is achieved by creating a
second game object with just a sprite renderer on the minimap's layer.
The main CP updates the alternate CP's position whenever it moves. The
scale isn't tied to minimap scale, so we'll have to tweak it if we
change the map size.

Various bug fixes. Just look through the changes, I don't want to write
them all.